refactor: split shared logic and fix realtime rendering issues
- Extracted shared functions to core.py (get_pose, fall_detection, etc.)
- Separated video.py and realtime.py for batch and live processing
- Fixed real-time display drawing misalignment by resizing frame to match letterbox
- Resolved real-time color distortion by ensuring proper RGB to BGR conversion
